further changes and tests
