--- src/app/[locale]/(logged-in)/user/profile/[tabId]/components/DetailsCV/index.ts
@@ -1,3 +1,0 @@
-import DetailsCV from "./DetailsCV";
-
-export default DetailsCV;
